flask_cbit_smart_attendance/requirements.txt
absl-py==0.15.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
arabic-reshaper==2.1.3
asgiref==3.3.4
astroid==2.4.2
astunparse==1.6.3
attrs==21.2.0
bcrypt==3.2.0
black==21.10b0
blinker==1.4
cachetools==4.2.4
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
clang==5.0
click==7.1.2
cloudpickle==2.0.0
colorama==0.4.3
cowsay==3.0
cycler==0.11.0
Cython==0.29.24
dataclasses==0.6
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-crispy-forms==1.11.2
django-debug-toolbar==1.8
django-environ==0.4.4
django-filter==2.4.0
django-heroku==0.3.1
django-multiselectfield==0.1.12
django-tinymce4-lite==1.8.0
django-widget-tweaks==1.4.1
djangorestframework==3.12.4
dnspython==2.0.0
email-validator==1.1.2
et-xmlfile==1.1.0
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
flatbuffers==1.12
future==0.18.2
fvcore==0.1.5.post20211023
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
gunicorn==20.0.4
h5py==3.1.0
html5lib==1.1
hydra-core==1.1.1
idna==2.10
imageio==2.10.1
imantics==0.1.12
imgaug==0.4.0
iopath==0.1.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsmin==2.2.2
jsonschema==4.1.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
labelme2coco==0.1.2
lazy-object-proxy==1.4.3
lxml==4.6.4
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.3
mccabe==0.6.1
migrate==0.3.7
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.19.5
oauthlib==3.1.1
omegaconf==2.1.1
opencv-contrib-python==4.5.4.58
opencv-python==4.5.4.58
openpyxl==3.0.9
opt-einsum==3.3.0
pathspec==0.9.0
Pillow==8.4.0
pipenv==2020.11.15
pixellib==0.7.1
platformdirs==2.4.0
portalocker==2.3.2
protobuf==3.19.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
PyJWT==1.7.1
pylint==2.5.3
PyMySQL==1.0.2
pyparsing==3.0.4
PyPDF2==1.26.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.18.0
pytesseract==0.3.8
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
pywin32==302
PyYAML==6.0
regex==2021.11.2
reportlab==3.5.68
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scipy==1.7.1
Shapely==1.8.0
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.4.1
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.1
tensorflow-estimator==2.6.0
termcolor==1.1.0
tifffile==2021.11.2
tk==0.1.0
toml==0.10.1
tomli==1.2.2
tqdm==4.62.3
twilio==6.59.0
typing-extensions==3.10.0.2
urllib3==1.26.2
virtualenv==20.1.0
virtualenv-clone==0.5.4
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
WTForms==2.3.3
xhtml2pdf==0.2.5
xlwt==1.3.0
xmljson==0.2.1
yacs==0.1.8
